# docs/swagger.yaml
definitions:
  main.InputJurosCompostos:
    properties:
      capital:
        type: number
      parcelas:
        type: integer
      taxa:
        type: number
    type: object
  main.InputSimulacaoParcela:
    properties:
      parcelas:
        type: integer
      taxa:
        type: number
      valor_produto:
        type: number
    type: object
  main.Montante:
    properties:
      mes:
        type: integer
      montante_acumulado:
        type: string
    type: object
  main.Simulacao:
    properties:
      amortizacao:
        type: string
      juros_pago:
        type: string
      mes:
        type: integer
      saldo_devedor:
        type: string
      valor_parcela:
        type: string
    type: object
info:
  contact: {}
  description: Exemplo de API com Swagger no Gin
  title: Juros Compostos API
  version: "1.0"
paths:
  /juros-compostos:
    post:
      consumes:
      - application/json
      description: Recebe valores de entrada e retorna o montante acumulado ao longo
        do tempo
      parameters:
      - description: Valores de entrada para o cálculo
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/main.InputJurosCompostos'
      produces:
      - application/json
      responses:
        "200":
          description: Lista de montantes calculados por mês
          schema:
            items:
              $ref: '#/definitions/main.Montante'
            type: array
        "400":
          description: Mensagem de erro para entrada inválida
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Calcula juros compostos
      tags:
      - Juros Compostos
  /ping:
    get:
      description: Retorna uma resposta "pong" indicando que a API está ativa
      produces:
      - application/json
      responses:
        "200":
          description: Resposta com mensagem 'pong'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verifica o status da API
      tags:
      - HealthCheck
  /simulacao-parcelas:
    post:
      consumes:
      - application/json
      description: Esta ação recebe os parâmetros do valor do produto, número de parcelas
        e taxa de juros, e retorna uma simulação das parcelas, com o valor da parcela,
        juros pagos, amortização e saldo devedor.
      parameters:
      - description: Dados de entrada para a simulação
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/main.InputSimulacaoParcela'
      produces:
      - application/json
      responses:
        "200":
          description: Simulação das parcelas com juros
          schema:
            items:
              $ref: '#/definitions/main.Simulacao'
            type: array
        "400":
          description: Erro de entrada com mensagem explicativa
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Simula o parcelamento de um produto com juros compostos.
      tags:
      - Simulação de Parcelas fixas em um financiamento com juros compostos
swagger: "2.0"